refactor(BrideMakerTest): ApplicationTest에 테스트 메서드 병합
